Add ability to identify Frama-C bool variable
Frama-C represents bools differently by prepending the name of the function
they belong in. The bool name and the function name are then separated by
a '_'. However, if we want to use it to create opaque predicate and insert
back into the source, that new variable name does not exist. So we have to
be able to identify bools, and then reconstruct their original names.